marble: sepolicy: Allow citsensorservice to write to persist
Xiaomi citsensorservice activates the "collision detection" sensor at
boot for whatever reason and keeps it always on. This sensor writes its
events to /mnt/vendor/persist/sensors/collisionData.txt. When its
unable to do so, it seems to enter an infinite loop which is consuming
the entire little cluster with CPU usage at 100%:

marblein:/sdcard # simpleperf stat -p 1896 --per-thread --duration 1
Performance counter statistics:

  mCollision-dete  1896  2936    1,867,233,051  cpu-cycles                # 1.798731 GHz
  als-algo         1896  2958       44,278,823  cpu-cycles                # 1.694063 GHz
  HwBinder:1896_1  1896  1972        5,874,148  cpu-cycles                # 1.478606 GHz
  citsensorservic  1896  2885        5,610,661  cpu-cycles                # 1.628860 GHz

Although we have no use for this sensor in AOSP, fix the underlying
issue first. We can remove or disable the sensor in a later change.